+++ app/src/main/java/com/zizi/rendezvous/FragmentListMeetings.java
@@ -23,8 +23,11 @@ import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.button.MaterialButton;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.Query;
@@ -52,10 +55,12 @@ public class FragmentListMeetings extends Fragment {
     private BadgeDrawable badgeDrawable; // для изменения количества непрочитанных сообщений
     private int countUnreads;
     private FragmentDetailsMeeting fragmentDetailsMeeting; // фрагмент с подробностями встречи
+    private DatabaseReference databaseReference;// ссылка на данные в БД
+    private FirebaseDatabase firebaseDatabase; // БД RealTime DataBase
 
     //вьюхи
     private BottomNavigationView bottomNavigationView; // нижняя панель с кнопками
-    private MaterialToolbar topAppBar; // верхняя панелька
+    private MaterialToolbar materialToolbar; // верхняя панелька
 
     public FragmentListMeetings() {
         // Required empty public constructor
@@ -101,6 +106,7 @@ public class FragmentListMeetings extends Fragment {
         //инициализация ////////////////////////////////////////////////////////////////////////////
         classGlobalApp = (ClassGlobalApp) getActivity().getApplicationContext();
         firebaseFirestore = FirebaseFirestore.getInstance(); //инициализация БД
+        firebaseDatabase = FirebaseDatabase.getInstance(); // БД
         userInfo = new HashMap<>(); // коллекция ключ-значение для описания встречи
         usersInfoAll = new ArrayList<>(); // информация по всем пользователям
         fragmentListChats = new FragmentListChats(); //фрагмент с чатами
@@ -113,8 +119,10 @@ public class FragmentListMeetings extends Fragment {
         recyclerView = getActivity().findViewById(R.id.rv_meeting); // список со встречами
         bottomNavigationView = getActivity().findViewById(R.id.bottom_navigation);
         activityMeetings = (ActivityMeetings)getActivity();
-        topAppBar = getActivity().findViewById(R.id.topAppBar);
-        //инициализация - КОНЕЦ
+        materialToolbar = getActivity().findViewById(R.id.materialToolbar);
+        //==========================================================================================
+
+
 
         // bottomNavigationView ////////////////////////////////////////////////////////////////////
         bottomNavigationView.setSelectedItemId(R.id.meetings); // делаем нужный пункт нижней панели по умолчанию
@@ -130,18 +138,36 @@ public class FragmentListMeetings extends Fragment {
             }
         });
 
-        ClassStaticMethods.getCountUnreads(bottomNavigationView); // подписываемся на обновление количества непрочитанных чатов на нижней панельке
-        //==========================================================================================
+        // ЗНАЧЕК с количеством непрочитанных сообщений текущего пользователя
+        databaseReference = firebaseDatabase.getReference("chats/unreads/" + classGlobalApp.GetCurrentUserUid() + "/");
+        databaseReference.addValueEventListener(new ValueEventListener() { // добавляем слушателя при изменении значения
+            @Override
+            public void onDataChange(@NonNull DataSnapshot snapshot) {
+                classGlobalApp.Log("FragmentListMeetings", "onActivityCreated/onDataChange", "Количество непрочитанных изменилось", false);
+                countUnreads = (int) snapshot.getChildrenCount(); // получаем количество непрочитанных чатов
+                if (countUnreads > 0) { // если есть непрочитанные чаты
+                    badgeDrawable = bottomNavigationView.getOrCreateBadge(R.id.chats); // создаем значек около вкладки Чаты на нижней панели, пока без номера
+                    badgeDrawable.setNumber(countUnreads); // показываем количество непрочитанных чатов
+                } else {
+                    bottomNavigationView.removeBadge(R.id.chats); // удаляем значек с панели
+                }
+            }
 
+            @Override
+            public void onCancelled(@NonNull DatabaseError error) {
+                // тут надо обработать ошибочку и залогировать
+            }
+        });
+        //============================================================================================
 
 
         // topAppBar ////////////////////////////////////////////////////////////////////////////////
-        topAppBar.setTitle("Встречи");
-        topAppBar.getMenu().findItem(R.id.request).setVisible(true); // показываем пункт заявки на встречу
-        topAppBar.setNavigationIcon(R.drawable.ic_outline_menu_24); // делаем кнопку навигации стрелочкой назад в верхней панельке
+        materialToolbar.setTitle("Встречи");
+        materialToolbar.getMenu().findItem(R.id.request).setVisible(true); // показываем пункт заявки на встречу
+        materialToolbar.setNavigationIcon(R.drawable.ic_outline_menu_24); // делаем кнопку навигации стрелочкой назад в верхней панельке
 
         // событие при клике на кнопку навигации на верхней панельке
-        topAppBar.setNavigationOnClickListener(new View.OnClickListener() {
+        materialToolbar.setNavigationOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
                 //getActivity().onBackPressed();
