+++ progress.txt
@@ -1,7 +1,4 @@
 НУЖНО СДЕЛАТЬ
-Удалить класс DataBase
-Удалить класс СтатикМетодс
-Проверить правильность работы максимального и минимального возраста
 63. Когда заходим на фрагмент чата, отметки о прочитанном делать в методе onResume
 65. Пересмотреть все жизненные циклы всех активити и фрагментов.
 59. Страничку в соц сети сделать ссылкой
